refactor: move hashLoginToken to core-server
