+++ imports/core-server/index.js
@@ -1,1 +1,2 @@
+export { default as hashLoginToken } from "./hashLoginToken";
 export * from "./registerFunction";
