+++ imports/plugins/core/cart/server/no-meteor/mutations/updateCartItemsQuantity.js
@@ -1,6 +1,6 @@
 import SimpleSchema from "simpl-schema";
 import ReactionError from "@reactioncommerce/reaction-error";
-import hashLoginToken from "/imports/plugins/core/accounts/server/no-meteor/util/hashLoginToken";
+import { hashLoginToken } from "/imports/core-server";
 
 const inputSchema = new SimpleSchema({
   "cartId": String,
