+++ imports/plugins/core/cart/server/no-meteor/util/getCartById.js
@@ -1,5 +1,5 @@
 import ReactionError from "@reactioncommerce/reaction-error";
-import hashLoginToken from "/imports/plugins/core/accounts/server/no-meteor/util/hashLoginToken";
+import { hashLoginToken } from "/imports/core-server";
 
 /**
  * @summary Gets a cart from the db by ID. If there is an account for the request, verifies that the
